feat: 헤더에 있던 로그아웃 버튼을 dashboard 페이지로 옮기고 ssr에 따른 on/off는 삭제
