+++ next/components/landing/Header/index.tsx
@@ -1,20 +1,35 @@
 import React from 'react';
+import Link from 'next/link';
 import { Box } from '@chakra-ui/react';
 
-import AuthButtons from '@components/landing/Header/AuthButtons';
-import { Props } from '@pages/index';
+import Button from '@components/common/Button';
 
 import * as S from './style';
 
 // TODO: 로그인/로그아웃 컴포넌트 작은 모니터에서 버거로 변경
 
-const Header = (props: Props) => (
+const Header = () => (
   <S.HeaderContainer>
     <Box w="fit-content">
       <S.MainLogo />
     </Box>
     <Box ml="auto">
-      <AuthButtons {...props} />
+      <S.ButtonsWrapper>
+        <Link href="/auth/sign-in">
+          <a>
+            <Button color="black" underline>
+              로그인
+            </Button>
+          </a>
+        </Link>
+        <Link href="/auth/sign-up">
+          <a>
+            <Button color="black" underline>
+              회원가입
+            </Button>
+          </a>
+        </Link>
+      </S.ButtonsWrapper>
     </Box>
   </S.HeaderContainer>
 );
