+++ next/components/sign-up/Optional/index.tsx
@@ -54,7 +54,7 @@ const Optional = ({ skillOptions, jobOptions }: IOptionalPageProps) => {
         localStorage.setItem('expiration', expiration);
         memoryStore.set(Memory.ACCESS_TOKEN, accessToken);
         window.sessionStorage.clear(); // 회원가입 과정에서 사용자가 입력했던 정보 삭제
-        router.replace('/service/dashboard');
+        router.replace('/service/');
       })
       .catch((err) => {
         setSignUpLoading(false);
